+++ src/main/java/com/team4/employeemood/repository/ProjectRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, Long> {
+public interface ProjectRepository extends JpaRepository<Project, Integer> {
 
-    Project findProjectById(Long projectId);
+    Project findProjectById(Integer projectId);
 
 }
